refactor: relocate agent_sudo from integrations to plugins
