+++ agent_runtimes/integrations/__init__.py
@@ -7,16 +7,13 @@ Integrations for agent-runtimes.
 This module provides integration with:
 - agent-codemode: Code-first MCP tool composition
 - agent-skills: Reusable agent skill management
-- Agent_Sudo: External tool authorization policy and audit integration
 """
 
-from .agent_sudo import authorize_tool_call
 from .codemode import CodemodeIntegration, get_codemode_integration
 from .tool_policy import evaluate_tool_request
 
 __all__ = [
     "CodemodeIntegration",
     "get_codemode_integration",
-    "authorize_tool_call",
     "evaluate_tool_request",
 ]
